upstream/0.0_git20150717

+++ test/defaultconflict/nc_test.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2013, Vastech SA (PTY) LTD. All rights reserved.
-// http://code.google.com/p/gogoprotobuf/gogoproto
+// http://github.com/gogo/protobuf/gogoproto
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are
@@ -29,16 +29,17 @@ package defaultcheck
 import (
 	"os"
 	"os/exec"
+	"strings"
 	"testing"
 )
 
 func testDefaultConflict(t *testing.T, name string) {
 	cmd := exec.Command("protoc", "--gogo_out=.", "-I=../../../../../:../../protobuf/:.", name+".proto")
 	data, err := cmd.CombinedOutput()
-	if err == nil {
-		t.Errorf("Expected error")
+	if err == nil && !strings.Contains(string(data), "Plugin failed with status code 1") {
+		t.Errorf("Expected error, got: %s", data)
 		if err := os.Remove(name + ".pb.go"); err != nil {
-			panic(err)
+			t.Error(err)
 		}
 	}
 	t.Logf("received expected error = %v and output = %v", err, string(data))
